fix(bbdo): replace index_id with uint64 type
* fix(conflict_manager_storage): replace index_data in uint64
* enh(status.hh): change type for index_data
* enh(storage:status): index_is is uint64 now.
* enh(bbdo/stream): add ulong in serialize and unserialize + unit test is ok
* enh(conflict_manager): replace index_id in uint64
* enh(conflict_manager): fix troncated index_id in metrics table
* enh(storage): fix syntax

REFS: MON-6801

+++ storage/inc/com/centreon/broker/storage/index_mapping.hh
@@ -41,7 +41,7 @@ namespace storage {
 class index_mapping : public io::data {
  public:
   index_mapping();
-  index_mapping(uint32_t index_id, uint32_t host_id, uint32_t service_id);
+  index_mapping(uint64_t index_id, uint32_t host_id, uint32_t service_id);
   index_mapping(index_mapping const& other) = delete;
   ~index_mapping() = default;
   index_mapping& operator=(index_mapping const& other) = delete;
@@ -50,7 +50,7 @@ class index_mapping : public io::data {
                                  storage::de_index_mapping>::value;
   }
 
-  uint32_t index_id;
+  uint64_t index_id;
   uint32_t host_id;
   uint32_t service_id;
